requirement.txt
numpy
Pillow==6.2.2
PyYAML==5.3.1
matplotlib==2.2.5
tqdm
scipy==1.2.3
tensorboardX==2.1
opencv-python